refactor(ui): extract task/agent tool renderer to tools-agent module
Slice 11b.1 step 11 of the message-part.tsx split. Pull the
task / agent card renderer — the only tool that owns navigation logic
(child session id resolution + clickable open) — into its own file.

New: packages/ui/src/components/message-part-tools-agent.tsx
  - renderAgentToolPart      shared by both "task" (legacy) and "agent"
                             ToolRegistry registrations.

Imports the routing helpers from `./message-part-session-link` rather
than reaching back into the aggregator.

tsgo --noEmit GREEN. bun test: 478 pass / 3 pre-existing fail.
