+++ packages/ui/src/components/message-part.tsx
@@ -207,87 +207,9 @@ export { MessageDivider }
 // extracted to `./message-part-tools-basic`. Side-effect import.
 import "./message-part-tools-basic"
 
-// Render function extracted so both "task" (legacy) and "agent" registrations share one reference.
-const renderAgentToolPart: ToolComponent = (props) => {
-  const data = useData()
-  const i18n = useI18n()
-  const location = useLocation()
-  const childSessionId = createMemo(() => {
-    const value = props.metadata.sessionId
-    if (typeof value === "string" && value) return value
-    return taskSession(props.input, location.pathname, data.store.session, data.store.agent)
-  })
-  const agent = createMemo(() => taskAgent(props.input.subagent_type, data.store.agent))
-  const title = createMemo(() => agent().name ?? i18n.t("ui.tool.agent.default"))
-  const tone = createMemo(() => agent().color)
-  const subtitle = createMemo(() => {
-    const value = props.input.description
-    if (typeof value === "string" && value) return value
-    return childSessionId()
-  })
-  const running = createMemo(() => props.status === "pending" || props.status === "running")
-
-  const href = createMemo(() => sessionLink(childSessionId(), location.pathname, data.sessionHref))
-  const clickable = createMemo(() => !!(childSessionId() && (data.navigateToSession || href())))
-
-  const open = () => {
-    const id = childSessionId()
-    if (!id) return
-    if (data.navigateToSession) {
-      data.navigateToSession(id)
-      return
-    }
-    const value = href()
-    if (value) window.location.assign(value)
-  }
-
-  const navigate = (event: MouseEvent) => {
-    if (!data.navigateToSession) return
-    if (event.button !== 0 || event.altKey || event.ctrlKey || event.metaKey || event.shiftKey) return
-    event.preventDefault()
-    open()
-  }
-
-  const trigger = () => (
-    <div data-component="task-tool-card">
-      <div data-slot="basic-tool-tool-info-structured">
-        <div data-slot="basic-tool-tool-info-main">
-          <Show when={running()}>
-            <span data-component="task-tool-spinner" style={{ color: tone() ?? "var(--brand-primary)" }}>
-              <Spinner />
-            </span>
-          </Show>
-          <span data-component="task-tool-title" style={{ color: tone() ?? "var(--fg-strong)" }}>
-            {title()}
-          </span>
-          <Show when={subtitle()}>
-            <span data-slot="basic-tool-tool-subtitle">{subtitle()}</span>
-          </Show>
-        </div>
-      </div>
-      <Show when={clickable()}>
-        <div data-component="task-tool-action">
-          <Icon name="square-arrow-top-right" />
-        </div>
-      </Show>
-    </div>
-  )
-
-  return (
-    <BasicTool
-      icon="agent"
-      status={props.status}
-      trigger={trigger()}
-      hideDetails
-      triggerHref={href()}
-      clickable={clickable()}
-      onTriggerClick={navigate}
-    />
-  )
-}
-
-ToolRegistry.register({ name: "task", render: renderAgentToolPart }) // agent-rename:legacy-render
-ToolRegistry.register({ name: "agent", render: renderAgentToolPart })
+// Slice 11b.1: task / agent renderer extracted to
+// `./message-part-tools-agent`. Side-effect import registers both names.
+import "./message-part-tools-agent"
 
 ToolRegistry.register({
   name: "bash",
